ScanGrad/requirements.txt
fastapi
uvicorn
httpx
pillow
pytesseract
anthropic
chromadb
sentence-transformers
python-dotenv
